refactor recent entries stuff to separate files
